requirement.txt
Flask
python-dotenv
langchain-community
langchain-google-genai
langchain
sentence-transformers
